Move extensions/pkg/util/chart to extensions/pkg/terraformer
