+++ extensions/pkg/terraformer/chart_test.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2019 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+// Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,22 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package chart
+package terraformer
 
 import (
 	"fmt"
-	"testing"
 
 	"github.com/gardener/gardener/pkg/chartrenderer"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"k8s.io/helm/pkg/manifest"
 )
-
-func TestUtil(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Chart Suite")
-}
 
 func mkManifest(name string, content string) manifest.Manifest {
 	return manifest.Manifest{
@@ -47,9 +42,9 @@ var _ = Describe("Chart", func() {
 
 			files, err := ExtractTerraformFiles(&chartrenderer.RenderedChart{
 				Manifests: []manifest.Manifest{
-					mkManifest(TerraformMainTFFilename, mainContent),
-					mkManifest(TerraformVariablesTFFilename, variablesContent),
-					mkManifest(TerraformTFVarsFilename, tfVarsContent),
+					mkManifest(MainKey, mainContent),
+					mkManifest(VariablesKey, variablesContent),
+					mkManifest(TFVarsKey, tfVarsContent),
 				},
 			})
